-added icon to ActiveCollectionsCard -fixed all functions names -changed step in ProgressBar component from 0.1% to 1% - rename constant.js to data.js
