+++ src/modules/mainPageCollection/components/ActiveCollectionsCard/ActiveCollectionsCard.module.scss
@@ -98,4 +98,13 @@
   font-weight: 400;
   font-size: 16px;
   line-height: 1.5;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+//icon style
+.collectionIcon {
+  width: 24px;
+  height: 24px;
 }
